eagle_vision/cv_service/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
supervision>=0.18.0
lap>=0.4.0
opencv-python>=4.8.0
kafka-python>=2.0.2
python-dotenv>=1.0.0
numpy>=1.24.0
